DeleteFileInPlaylist: Only remove from list after confirming file delete/recycle

+++ src/mpc-hc/PlayerPlaylistBar.cpp
@@ -1376,28 +1376,29 @@ bool CPlayerPlaylistBar::DeleteFileInPlaylist(POSITION pos, bool recycle)
     }
 
     CString filename = m_pl.GetAt(pos).m_fns.GetHead();
+    int listPos = FindItem(pos);
 
     // Get position of next file
     POSITION nextpos = pos;
     if (isplaying) {
         m_pl.GetNext(nextpos);
     }
-    // remove selected from playlist
-    int listPos = FindItem(pos);
-    m_list.DeleteItem(listPos);
-    m_list.RedrawItems(listPos, m_list.GetItemCount() - 1);
-    m_pl.RemoveAt(pos);
-    SavePlaylist();
 
     // Delete file
-    FileDelete(filename, m_pMainFrame->m_hWnd, recycle);
-    // Continue with next file
-    if (isplaying) {
-        if (folderPlayNext) {
-            m_pMainFrame->DoAfterPlaybackEvent(); //we know this will call PLAY_NEXT, which should do normal folder looping
-        } else if (nextpos || AfxGetAppSettings().bLoopFolderOnPlayNextFile) {
-            m_pl.SetPos(nextpos);
-            m_pMainFrame->OpenCurPlaylistItem();
+    if (SUCCEEDED(FileDelete(filename, m_pMainFrame->m_hWnd, recycle))) {
+        // remove selected from playlist
+        m_list.DeleteItem(listPos);
+        m_list.RedrawItems(listPos, m_list.GetItemCount() - 1);
+        m_pl.RemoveAt(pos);
+        SavePlaylist();
+        // Continue with next file
+        if (isplaying) {
+            if (folderPlayNext) {
+                m_pMainFrame->DoAfterPlaybackEvent(); //we know this will call PLAY_NEXT, which should do normal folder looping
+            } else if (nextpos || AfxGetAppSettings().bLoopFolderOnPlayNextFile) {
+                m_pl.SetPos(nextpos);
+                m_pMainFrame->OpenCurPlaylistItem();
+            }
         }
     }
